backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# AI & ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
faiss-cpu==1.12.0
tiktoken==0.5.2

# Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
email-validator==2.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Data Processing
pandas>=2.2.1
numpy>=1.26.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# NLP
spacy==3.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS
python-cors==1.0.0
